#781 [REFACTOR] 시험후기 리스트 페이지와 상세 페이지 폴더를 분리
