+++ src/pages/ExamReviewPage/index.js
@@ -1,4 +1,1 @@
-export { default as ExamReviewPage } from './ExamReviewPage/ExamReviewPage';
-export { default as ExamReviewDetailPage } from './ExamReviewDetailPage/ExamReviewDetailPage';
-export { default as ExamReviews } from './ExamReviews/ExamReviews';
-export { default as ExamReviewSearchList } from './ExamReviewSearchList/ExamReviewSearchList';
+export { default as ExamReviewPage } from './ExamReviewPage';
