+++ src/pages/ExamReviewsPage/components/ExamReviews/ExamReviews.module.css
@@ -1,0 +1,16 @@
+.list {
+  padding: 0.25rem 1rem 6rem 1rem;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 0.5rem;
+  overflow: hidden;
+}
+
+.to {
+  width: 100%;
+}
+
+.loading {
+  padding-bottom: 7rem;
+}
