+++ src/pages/ExamReviewsPage/index.js
@@ -1,0 +1,4 @@
+export { default as ExamReviewsPage } from './ExamReviewsPage';
+
+export { default as ExamReviews } from './components/ExamReviews/ExamReviews';
+export { default as ExamReviewSearchList } from './components/ExamReviewSearchList/ExamReviewSearchList';
